refactor(whatsrust): relocate forwarding result types
